test: add fair timer
